feat(publisher-dashboard): refactor dashboard components and enhance data visualization
- Introduced new components: `AnalyticsGrid`, `DailyClicksChart`, `DeviceDistributionChart`, and `RevenueChart` for improved data representation.
- Streamlined the `PublisherDashboard` by removing unused imports and mock data, enhancing maintainability.
- Implemented complete data handling for daily clicks and revenue charts, ensuring accurate visualizations.
- Enhanced loading states with skeleton components for better user experience during data fetching.
- Updated the layout to utilize a grid system for better responsiveness and organization of dashboard elements.
