+++ src/components/publisher/dashboard/_components/revenue-chart.tsx
@@ -1,0 +1,160 @@
+import { eachDayOfInterval, format } from "date-fns"
+import { Info } from "lucide-react"
+import { CartesianGrid, Line, LineChart, XAxis, YAxis } from "recharts"
+
+import { formatVNDCurrency } from "@/lib/utils"
+
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
+import {
+  ChartContainer,
+  ChartLegend,
+  ChartLegendContent,
+  ChartTooltip,
+} from "@/components/ui/chart"
+import { Skeleton } from "@/components/ui/skeleton"
+import {
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+  Tooltip as UITooltip,
+} from "@/components/ui/tooltip"
+
+interface RevenueData {
+  date: string
+  revenue: number
+}
+
+interface RevenueChartProps {
+  isLoading: boolean
+  data: RevenueData[]
+  startDate: Date
+  endDate: Date
+}
+
+export function RevenueChart({
+  isLoading,
+  data,
+  startDate,
+  endDate,
+}: RevenueChartProps) {
+  // Generate array of all dates in the range
+  const dateRange = eachDayOfInterval({
+    start: startDate,
+    end: endDate,
+  })
+
+  // Create a map of existing data
+  const dataMap = new Map(data.map((item) => [item.date, item]))
+
+  // Generate complete dataset with 0s for missing dates
+  const completeData = dateRange.map((date) => {
+    const formattedDate = format(date, "MMM d")
+    const existingData = dataMap.get(formattedDate)
+
+    return {
+      date: formattedDate,
+      revenue: existingData?.revenue ?? 0,
+    }
+  })
+
+  return (
+    <Card className="col-span-4">
+      <CardHeader className="mb-5 border-b px-6 py-3">
+        <div className="flex items-center justify-between">
+          <div className="flex items-center gap-3">
+            <CardTitle className="text-base font-medium">
+              Revenue Overview
+            </CardTitle>
+            <TooltipProvider>
+              <UITooltip>
+                <TooltipTrigger>
+                  <Info className="size-4 text-muted-foreground/60" />
+                </TooltipTrigger>
+                <TooltipContent>
+                  Daily revenue from verified clicks
+                </TooltipContent>
+              </UITooltip>
+            </TooltipProvider>
+          </div>
+        </div>
+      </CardHeader>
+      <CardContent className="justify-end">
+        {isLoading ? (
+          <Skeleton className="h-[320px] w-full" />
+        ) : !completeData || completeData.length === 0 ? (
+          <div className="flex h-[320px] w-full flex-col items-center justify-center rounded-lg border-2 border-dashed text-center">
+            <h3 className="text-lg font-semibold">No Revenue Data Available</h3>
+            <p className="mt-2 text-sm text-muted-foreground">
+              No revenue recorded for the selected period
+            </p>
+          </div>
+        ) : (
+          <ChartContainer
+            className="h-[320px] w-full"
+            config={{
+              revenue: {
+                label: "Revenue",
+                color: "#10b981",
+              },
+            }}
+          >
+            <LineChart
+              accessibilityLayer
+              data={completeData}
+              margin={{ top: 20 }}
+            >
+              <CartesianGrid
+                strokeDasharray="3 3"
+                vertical={false}
+                stroke="#f1f5f9"
+              />
+              <XAxis
+                dataKey="date"
+                tick={{ fontSize: 12 }}
+                tickLine={false}
+                axisLine={false}
+                stroke="#94a3b8"
+              />
+              <YAxis
+                tick={{ fontSize: 12 }}
+                tickLine={false}
+                axisLine={false}
+                stroke="#94a3b8"
+                tickFormatter={(value) => formatVNDCurrency(value)}
+              />
+              <ChartTooltip
+                content={({ active, payload }) => {
+                  if (!active || !payload?.length) return null
+                  const value = payload[0]?.value
+                  if (typeof value !== "number") return null
+                  return (
+                    <div className="rounded-lg border bg-white p-2 shadow-sm">
+                      <div className="font-medium">
+                        {payload[0].payload.date}
+                      </div>
+                      <div className="mt-1 flex items-center gap-2">
+                        <div className="size-2 rounded-full bg-emerald-500" />
+                        <span className="text-sm text-muted-foreground">
+                          {formatVNDCurrency(value)}
+                        </span>
+                      </div>
+                    </div>
+                  )
+                }}
+              />
+              <ChartLegend content={<ChartLegendContent />} />
+
+              <Line
+                type="linear"
+                dataKey="revenue"
+                stroke="#10b981"
+                strokeWidth={2}
+                dot={false}
+              />
+            </LineChart>
+          </ChartContainer>
+        )}
+      </CardContent>
+    </Card>
+  )
+}
